Remove CoreFunctions.php file, as PHPStan no longer complains about having all helper functions in Core.php
